Resolves review remarks
