--- basyx.submodelrepository/basyx.submodelrepository-core/src/test/java/org/eclipse/digitaltwin/basyx/submodelrepository/core/util/TestSubmodelRepositoryUtilities.java
@@ -1,63 +1,0 @@
-/*******************************************************************************
- * Copyright (C) 2023 the Eclipse BaSyx Authors
- *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
- * 
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
- * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
- * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- * SPDX-License-Identifier: MIT
- ******************************************************************************/
-package org.eclipse.digitaltwin.basyx.submodelrepository.core.util;
-
-import java.util.Arrays;
-import java.util.Collection;
-
-import org.eclipse.digitaltwin.aas4j.v3.model.Submodel;
-import org.eclipse.digitaltwin.basyx.core.exceptions.CollidingIdentifierException;
-import org.eclipse.digitaltwin.basyx.submodelrepository.util.SubmodelRepositoryUtilities;
-import org.eclipse.digitaltwin.basyx.submodelservice.DummySubmodelFactory;
-import org.junit.Test;
-import org.junit.Test.None;
-
-/**
- * 
- * @author jungjan
- *
- */
-public class TestSubmodelRepositoryUtilities {
-
-	@Test(expected = CollidingIdentifierException.class)
-	public void assertIdUniquenessColliding() {
-		Collection<Submodel> submodelsWithCollidingIds = createSubmodelCollectionWithCollidingIds();
-		SubmodelRepositoryUtilities.assertIdUniqueness(submodelsWithCollidingIds);
-	}
-
-	@Test(expected = None.class)
-	public void assertIdUniquenessUniquess() {
-		Collection<Submodel> submodelsWithUniqueIds = createSubmodelCollectionWithUniqueIds();
-		SubmodelRepositoryUtilities.assertIdUniqueness(submodelsWithUniqueIds);
-	}
-
-	private Collection<Submodel> createSubmodelCollectionWithCollidingIds() {
-		return Arrays.asList(DummySubmodelFactory.createTechnicalDataSubmodel(), DummySubmodelFactory.createTechnicalDataSubmodel());
-	}
-
-	private Collection<Submodel> createSubmodelCollectionWithUniqueIds() {
-		return Arrays.asList(DummySubmodelFactory.createSimpleDataSubmodel(), DummySubmodelFactory.createTechnicalDataSubmodel());
-	}
-}
